@@ -1,22 +1,5 @@
-const getFullYear = () => {
-  let current_year = new Date().getFullYear();
-  return current_year;
-}
+export const getFullYear = () => new Date().getFullYear();
 
-const getFooterCopy = (isIndex) => {
-  if (isIndex === true) {
-    return "Holberton School";
-  } else {
-    return "Holberton School main dashboard";
-  }
-}
+export const getFooterCopy = (isIndex) => (isIndex ? "Holberton School" : "Holberton School main dashboard");
 
-const getLatestNotification = () => {
-  return "<strong>Urgent requirement</strong> - complete by EOD";
-}
-
-module.exports = {
-  getFooterCopy,
-  getFullYear,
-  getLatestNotification
-};
+export const getLatestNotification = () => "<strong>Urgent Requirement</strong> - complete by EOD";
