@@ -1,10 +1,10 @@
-import React, { Component } from 'react'
-import PropTypes from 'prop-types';
+import PropTypes from "prop-types";
+import React, { Component } from "react";
 
 class BodySection extends Component {
   render() {
     return (
-      <div className='bodySection'>
+      <div className="bodySection">
         <h2>{this.props.title}</h2>
         {this.props.children}
       </div>
@@ -12,16 +12,9 @@ class BodySection extends Component {
   }
 }
 
-BodySection. defaultProps = {
-	children: <React.Fragment />
-};
-
 BodySection.propTypes = {
   title: PropTypes.string.isRequired,
-  children: PropTypes.oneOfType([
-    PropTypes.string,
-    PropTypes.element
-  ])
+  children: PropTypes.oneOfType([PropTypes.string, PropTypes.element]),
 };
 
 export default BodySection;
