@@ -1,28 +1,27 @@
-import React, { Component } from 'react'
-import BodySection from './BodySection';
-import './BodySectionWithMarginBottom.css';
-import PropTypes from 'prop-types';
+import PropTypes from "prop-types";
+import React, { Component } from "react";
+import BodySection from "./BodySection";
+import { StyleSheet, css } from "aphrodite";
 
 class BodySectionWithMarginBottom extends Component {
   render() {
     return (
-      <div className='bodySectionWithMargin'>
-        <BodySection {...this.props}/>
+      <div className={css(styles.bodySectionWithMargin)}>
+        <BodySection {...this.props} />
       </div>
     );
   }
 }
 
-BodySectionWithMarginBottom. defaultProps = {
-	children: <React.Fragment />
-};
+const styles = StyleSheet.create({
+  bodySectionWithMargin: {
+    marginBottom: "40",
+  },
+});
 
 BodySectionWithMarginBottom.propTypes = {
   title: PropTypes.string.isRequired,
-  children: PropTypes.oneOfType([
-    PropTypes.string,
-    PropTypes.element
-  ])
+  children: PropTypes.oneOfType([PropTypes.string, PropTypes.element]),
 };
 
 export default BodySectionWithMarginBottom;
