@@ -1,12 +1,21 @@
 import React from "react";
-import { shallow } from 'enzyme';
-import BodySectionWithMarginBottom from './BodySectionWithMarginBottom';
+import { shallow } from "enzyme";
+import BodySectionWithMarginBottom from "./BodySectionWithMarginBottom";
+import BodySection from "./BodySection";
+import { StyleSheetTestUtils } from "aphrodite";
 
-describe("Testing BodySectionWithMarginBottom Component",() => {
-  it('checking that shallowing the component should render correctly a BodySection component and that the props are passed correctly to the child component', () => {
-    let wrapper = shallow(<BodySectionWithMarginBottom title="test title" />);
-		expect(wrapper.find("BodySection").exists()).toBe(true);
-    wrapper = shallow(<BodySectionWithMarginBottom title="test title"><p>test children</p></BodySectionWithMarginBottom>)
-		expect(wrapper.find("BodySection").props().title).toBe('test title');
+beforeEach(() => {
+  StyleSheetTestUtils.suppressStyleInjection();
+});
+afterEach(() => {
+  StyleSheetTestUtils.clearBufferAndResumeStyleInjection();
+});
+
+describe("BodySectionWithMarginBottom tests", () => {
+  it("should apply margin bottom to child component", () => {
+    const wrapper = shallow(<BodySectionWithMarginBottom title="test title" />);
+
+    expect(wrapper.find(BodySection)).toHaveLength(1);
+    expect(wrapper.find(BodySection).html()).toEqual('<div class="bodySection"><h2>test title</h2></div>');
   });
 });
