@@ -1,8 +1,9 @@
-import PropTypes from 'prop-types';
+import PropTypes from "prop-types";
 
-export default PropTypes.shape({
+const CourseShape = PropTypes.shape({
   id: PropTypes.number.isRequired,
   name: PropTypes.string.isRequired,
-  credit: PropTypes.number.isRequired
-}
-);
+  credit: PropTypes.number.isRequired,
+});
+
+export default CourseShape;
