@@ -1,13 +1,15 @@
-import React from 'react';
-import './Footer.css';
-import { getFooterCopy, getFullYear } from '../utils/utils';
+import React from "react";
+import "./Footer.css";
+import { getFullYear, getFooterCopy } from "../utils/utils";
 
 function Footer() {
   return (
-    <div className="App-footer">
-      <p>Copyright {getFullYear()} - {getFooterCopy()}</p>
-    </div>
-  )
+    <>
+      <div className="App-footer">
+        Copyright {getFullYear()} - {getFooterCopy()}
+      </div>
+    </>
+  );
 }
 
 export default Footer;
