@@ -1,24 +1,23 @@
-import React from 'react';
-import Header from './Header';
-import { shallow } from 'enzyme';
+import React from "react";
+import Header from "./Header";
+import { shallow } from "enzyme";
+import { StyleSheetTestUtils } from "aphrodite";
 
+beforeEach(() => {
+  StyleSheetTestUtils.suppressStyleInjection();
+});
+afterEach(() => {
+  StyleSheetTestUtils.clearBufferAndResumeStyleInjection();
+});
 
-describe('Testing <Header /> Component', () => {
-  let wrapper;
-
-  beforeEach(() => {
-    wrapper = shallow(<Header />);
+describe("Header", () => {
+  it("render without crashing", () => {
+    const wrapper = shallow(<Header />);
+    expect(wrapper.exists()).toEqual(true);
   });
-
-  it("Renders with out crashing", () => {
-    expect(wrapper).toBeDefined();
-  });
-
-  it("Render an h1 tag", () => {
-    expect(wrapper.find('h1')).toBeDefined();
-  });
-
-  it("Render an img tag", () => {
-    expect(wrapper.find('img')).toBeDefined();
+  it("should render a h1", () => {
+    const wrapper = shallow(<Header />);
+    expect(wrapper.exists("img")).toEqual(true);
+    expect(wrapper.containsMatchingElement(<h1>School dashboard</h1>)).toEqual(true);
   });
 });
