@@ -1,19 +1,37 @@
-import React from 'react';
-import './Login.css';
+import React from "react";
+import { StyleSheet, css } from "aphrodite";
 
 function Login() {
   return (
     <React.Fragment>
-      <div className="App-body">
+      <div className={css(styles["App-body"])}>
         <p>Login to access the full dashboard</p>
-        <label htmlFor="email">Email: </label>
-        <input type="email" id="email" name="email" />
-        <label htmlFor="password">Password: </label>
-        <input type="password" id="password" name="password" />
-        <button>OK</button>
+        <form>
+          <label htmlFor="email">Email:</label>
+          <input className={css(styles.input)} type="email" name="email"></input>
+          <label htmlFor="password">Password:</label>
+          <input className={css(styles.input)} type="password" name="password"></input>
+          <button>OK</button>
+        </form>
       </div>
     </React.Fragment>
-  )
+  );
 }
+
+const styles = StyleSheet.create({
+  "App-body": {
+    fontSize: "1rem",
+    padding: "2em",
+    height: "45%",
+    "@media (max-width: 900px)": {
+      display: "flex",
+      flexDirection: "column",
+    },
+  },
+
+  input: {
+    margin: "10px",
+  },
+});
 
 export default Login;
