@@ -1,23 +1,23 @@
-import React from 'react';
-import Login from './Login';
-import { shallow } from 'enzyme';
+import { shallow } from "enzyme";
+import React from "react";
+import Login from "./Login";
+import { StyleSheetTestUtils } from "aphrodite";
 
-describe("testing the <Login /> component", () => {
-  let wrapper;
+beforeEach(() => {
+  StyleSheetTestUtils.suppressStyleInjection();
+});
+afterEach(() => {
+  StyleSheetTestUtils.clearBufferAndResumeStyleInjection();
+});
 
-  beforeEach(() => {
-    wrapper = shallow(<Login />);
+describe("Header", () => {
+  it("should render without crashing", () => {
+    const wrapper = shallow(<Login />);
+    expect(wrapper.exists()).toEqual(true);
   });
-
-  it("Login component renders without crashing", () => {
-    expect(wrapper).toBeDefined();
-  });
-
-  it("Login component renders 2 input tags", () => {
-    expect(wrapper.find("input")).toHaveLength(2);
-  });
-
-  it("Login component renders 2 label tags", () => {
+  it("should have 2 input tags and 2 label tags", () => {
+    const wrapper = shallow(<Login />);
     expect(wrapper.find("label")).toHaveLength(2);
+    expect(wrapper.find("input")).toHaveLength(2);
   });
 });
