@@ -1,19 +1,19 @@
-import React from 'react';
-import closeIcon from '../assets/close-icon.png';
-import NotificationItem from './NotificationItem';
-import PropTypes from 'prop-types';
-import NotificationItemShape from './NotificationItemShape';
-import { StyleSheet, css } from 'aphrodite';
+import React, { Component } from "react";
+import { StyleSheet, css } from "aphrodite";
+import closeIcon from "../assets/close-icon.png";
+import NotificationItem from "./NotificationItem";
+import PropTypes from "prop-types";
+import NotificationItemShape from "./NotificationItemShape";
 
-class Notifications extends React.Component {
+class Notifications extends Component {
   constructor(props) {
     super(props);
+
     this.markAsRead = this.markAsRead.bind(this);
   }
 
   shouldComponentUpdate(nextProps) {
-    return nextProps.listNotifications.length > this.props.listNotifications.length ||
-    this.props.displayDrawer != nextProps.displayDrawer;
+    return nextProps.length > this.props.listNotifications.length;
   }
 
   markAsRead(id) {
@@ -22,134 +22,121 @@ class Notifications extends React.Component {
 
   render() {
     return (
-      <>
-        {!this.props.displayDrawer ?
-          <div className={css(notificationStyles.menuItem)} onClick={this.props.handleDisplayDrawer}>
-            Your notifications
+      <React.Fragment>
+        {!this.props.displayDrawer ? (
+          <div className={css(styles.menuItem)} onClick={this.props.handleDisplayDrawer}>
+            <p>Your notifications</p>
           </div>
-        :
-          <div className={css(notificationStyles.notifications)}>
-            <button style={{
-              color: '#3a3a3a',
-              fontWeight: 'bold',
-              background: 'none',
-              border: 'none',
-              fontSize: '15px',
-              position: 'absolute',
-              right: '3px',
-              top: '3px',
-              cursor: 'pointer',
-              outline: 'none',
-            }}
-            aria-label="Close"
-            className={css(notificationStyles.button)}
-            onClick={(e) => {
-              console.log('Close button has been clicked');
-              this.props.handleHideDrawer();
-            }}
+        ) : (
+          <div className={css(styles.Notifications)}>
+            <button
+              style={{
+                color: "#3a3a3a",
+                fontWeight: "bold",
+                background: "none",
+                border: "none",
+                fontSize: "15px",
+                position: "absolute",
+                right: "3px",
+                top: "3px",
+                cursor: "pointer",
+                outline: "none",
+              }}
+              aria-label="Close"
+              onClick={(e) => {
+                console.log("Close button has been clicked");
+                this.props.handleHideDrawer();
+              }}
             >
-              <img src={closeIcon} alt="close icon" width="15px" />
+              <img src={closeIcon} alt="close icon" width="10px" />
             </button>
-            {
-              this.props.listNotifications.length != 0 ?
-                <p>Here is the list of notifications</p>
-              : null
-            }
-            <ul className={css(notificationStyles.ul)}>
-              {
-                this.props.listNotifications.length == 0 ?
-                  <NotificationItem type="default" value="No new notification for now" />
-                : null
-              }
-              {
-                this.props.listNotifications.map((val, idx)=> {
-                  return <NotificationItem
-                  type={val.type}
-                  value={val.value}
-                  html={val.html}
-                  key={val.id}
-                  markAsRead={this.markAsRead}
-                  id={val.id}
-                />
-                })
-              }
+            {this.props.listNotifications.length != 0 ? <p>Here is the list of notifications</p> : null}
+            <ul>
+              {this.props.listNotifications.length == 0 ? <NotificationItem type="default" value="No new notification for now" /> : null}
+              {this.props.listNotifications.map((val, idx) => {
+                return <NotificationItem type={val.type} value={val.value} html={val.html} key={val.id} markAsRead={this.markAsRead} id={val.id} />;
+              })}
             </ul>
           </div>
-        }
-        
-      </>
+        )}
+      </React.Fragment>
     );
   }
 }
 
 const opacityAnim = {
-  '0%': { opacity: 0.5 },
-  '100%': { opacity: 1}
+  "0%": { opacity: 0.5 },
+  "100%": { opacity: 1 },
 };
 
 const bounceAnim = {
-  '0%': { transform: 'translateY(0px)' },
-  '33%': { transform: 'translateY(-5px)'},
-  '66%': { transform: 'translateY(5px)'},
-  '100%': { transform: 'translateY(0px)'},
+  "0%": { transform: "translateY(0px)" },
+  "33%": { transform: "translateY(-5px)" },
+  "66%": { transform: "translateY(5px)" },
+  "100%": { transform: "translateY(0px)" },
 };
 
-const notificationStyles = StyleSheet.create({
-	notifications: {
-    border: '3px dotted var(--holberton-red)',
-    padding: '6px 12px',
-    position: 'absolute',
-    top: '21px',
-    right: '7px',
-    marginTop: '12px',
-    zIndex: '100',
-    '@media (max-width: 900px)': {
-      width: '100%',
-      padding: '0px',
+const styles = StyleSheet.create({
+  Notifications: {
+    padding: "1em",
+    border: "2px dashed red",
+    position: "absolute",
+    top: "1.8em",
+    right: "0",
+    zIndex: "100",
+    "@media (max-width: 900px)": {
+      width: "100%",
+      padding: "0px",
       fontSize: 20,
-      position: 'relative',
+      position: "relative",
       right: 0,
       left: 0,
-      border: 'none',
-    }
-	},
-  menuItem: {
-    position: 'relative',
-    zIndex: 100,
-    float: 'right',
-    backgroundColor: '#fff8f8',
-    ':hover': {
-      cursor: 'pointer',
-      animationName: [opacityAnim, bounceAnim],
-      animationDuration: '1s, 0.5s',
-      animationIterationCount: '3'
-    }
+      border: "none",
+    },
   },
+
+  "notification-header": {
+    display: "flex",
+    justifyContent: "space-between",
+  },
+
+  menuItem: {
+    position: "relative",
+    zIndex: 100,
+    textAlign: "right",
+    ":hover": {
+      cursor: "pointer",
+      animationName: [opacityAnim, bounceAnim],
+      animationDuration: "1s, 0.5s",
+      animationIterationCount: "3",
+    },
+  },
+
   ul: {
-    '@media (max-width: 900px)': {
-      padding: 0
-    }
+    "@media (max-width: 900px)": {
+      padding: 0,
+    },
   },
   button: {
-    '@media (max-width: 900px)': {
-      position: 'relative',
-      float: 'right',
-    }
-  }
+    "@media (max-width: 900px)": {
+      position: "relative",
+      float: "right",
+    },
+  },
 });
-
-Notifications.defaultProps = {
-  displayDrawer: false,
-  listNotifications: [],
-  handleHideDrawer: () => {},
-  handleDisplayDrawer: () => {}
-};
 
 Notifications.propTypes = {
   displayDrawer: PropTypes.bool,
   listNotifications: PropTypes.arrayOf(NotificationItemShape),
+  handleDisplayDrawer: PropTypes.func,
   handleHideDrawer: PropTypes.func,
-  handleDisplayDrawer: PropTypes.func
+};
+
+Notifications.defaultProps = {
+  displayDrawer: false,
+  listNotifications: [],
+  handleDisplayDrawer: () => {},
+  handleHideDrawer: () => {},
 };
 
 export default Notifications;
